Sliders are Working!
Managed to get an old module working. The Vueperslides node module was acting up with an 'undefined window' error. Turns out it was because of the css file it needed not be loaded correctly (or something...).

I made the vueperslides components a plugin file, and used an if statement to check if the process.client variable to load in the css files correctly. I'm not entirely sure why this works, it just does, and nothing to my knowledge has broke because of it.

I'm gonna mess around with it a bit more until I'm satisfied that it has been implemented cleanly. If anyone feels like styling it, there's documentation online with many examples for making it look much nicer. Also, since it is a global component (via plugins), I can suggest other to take a look at it as well, if it is suitable for other pages.

Unfortunately it is not server-side renderable, but it might be able to show new documentation as it is added. No guarantees, however.

Final Notes: updated _documents index.vue, updated the nuxt.config.js file for the slides, removed the vue-carousel component and its associated information in both package.json files, created vueperslides.js plugin file.