Simple cleanup
Did some cleanup. I noticed AppSection's title prop. was causing issues, in that a title (specified or not) would be displayed, making the diagram page look rather awkward. I removed the prop. and kept the h2 CSS styling so any h2 headings in the <slot /> part of an AppSection can still receive that styling. Since then, I've fixed up the home page and the folders page to display their titles in their respective AppSection slots.

The diagram-page2 page now uses the AppSection version from component folder 'Utils'.

I've removed AppVideo2 and modified AppVideo. AppVideo now displays Youtube videos.

A better placeholder video has been placed on the homepage, relevant to ontologies.

Found a sneaky </p> that shouldn't have existed in the homepage.

Changed the AppModal component in the _documents page to the version of AppModal in the 'Utilities' component folder (was the version in 'Utils').