Test Cases and Documentation
Test cases A, B, D and E have been completed and edited.

Documentation for Selenium and Testing have been done up. Further additions to the .md files are welcome.

Additional naming conventions added to several components. There shouldn't be any issues as a result of it due to being unobtrusive to the rest of the code.

Bug reports is testing logs have accompanying images of said bugs.